Error

+++ src/Main.java
@@ -3,7 +3,7 @@ public class Main {
         BmiService service = new BmiService();
         int weight = 100;
         float length = 1.72F;
-        float index = service.calculate(weight, length);
+        double index = service.calculate(weight, length);
         System.out.println(index);
     }
 }
